requirements.txt
# Core AI/ML Dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langsmith>=0.0.60
langserve>=0.0.30

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Data Processing
pydantic>=2.5.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# Vector Database
chromadb>=0.4.0

# Database
sqlalchemy>=2.0.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0

# Logging and Monitoring
python-json-logger>=2.0.7

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0